requirements.txt
streamlit
pillow
pytesseract
numpy
opencv-python-headless
